Plot updates
In this commit, I added some wrangling to help with the plotting of the data. Initially the scales of the visualizations were hard to interpret, so wrangling the data helped fix that. I also included all the plots required for each of the predictor variables along with a small comment providing an initial observational analysis for each of the plots.